--- BlazarUI/app/scripts/actions/newBranchActions.js
@@ -1,13 +1,0 @@
-/*global config*/
-import Reflux from 'reflux';
-
-const NewBranchActions = Reflux.createActions([
-  'loadBranchBuildHistory',
-  'loadBranchInfo',
-  'loadBranchModules',
-  'loadBranchMalformedFiles',
-  'startPolling',
-  'stopPolling'
-]);
-
-export default NewBranchActions;
