--- BlazarUI/app/scripts/actions/newRepoBuildActions.js
@@ -1,11 +1,0 @@
-/*global config*/
-import Reflux from 'reflux';
-
-const NewRepoBuildActions = Reflux.createActions([
-  'loadRepoBuild',
-  'loadModuleBuilds',
-  'startPolling',
-  'stopPolling'
-]);
-
-export default NewRepoBuildActions;
