+++ BlazarUI/app/scripts/data/RepoBuildApi.js
@@ -1,65 +1,91 @@
 /*global config*/
-import RepoBuildPollingProvider from '../services/RepoBuildPollingProvider';
 import Resource from '../services/ResourceProvider';
+import Q from 'q';
+import { findWhere, map, extend } from 'underscore';
+import humanizeDuration from 'humanize-duration';
 
-class RepoBuildApi {
-
-  constructor(params) {
-    this.buildsPoller = new RepoBuildPollingProvider(params);
+function _parse(resp) {
+  if (resp.startTimestamp && resp.endTimestamp) {
+    resp.duration = humanizeDuration(resp.endTimestamp - resp.startTimestamp, {round: true});
   }
 
-  startPolling(cb) {
-    this.cb = cb;
+  return resp;
+}
 
-    this.buildsPoller.poll((err, resp) => {
-      if (err) {
-        return cb(err);
-      }
+function _fetchModuleNames(params) {
+  const moduleNamesPromise = new Resource({
+    url: `${config.apiRoot}/branches/${params.branchId}/modules`,
+    type: 'GET'
+  }).send();
 
-      this.branchId = resp.branchId;
-      cb(err, resp);
+  return moduleNamesPromise;
+}
+
+function _fetchBranchBuildHistory(params) {
+  const branchBuildHistoryPromise = new Resource({
+    url: `${config.apiRoot}/builds/history/branch/${params.branchId}`,
+    type: 'GET'
+  }).send();
+
+  return branchBuildHistoryPromise;
+}
+
+function fetchRepoBuild(params, cb) {
+  return _fetchBranchBuildHistory(params).then((resp) => {
+    const repoBuild = findWhere(resp, {buildNumber: parseInt(params.buildNumber, 10)});
+    const repoBuildPromise = new Resource({
+      url: `${config.apiRoot}/branches/builds/${repoBuild.id}`,
+      type: 'GET'
+    }).send();
+
+    return repoBuildPromise.then((resp) => {
+      return cb(_parse(resp));
     });
-  }
+  });
+}
 
-  stopPolling() {
-    if (!this.buildsPoller) {
-      return;
-    }
+function fetchModuleBuilds(params, cb) {
+  return _fetchBranchBuildHistory(params).then((resp) => {
+    const repoBuild = findWhere(resp, {buildNumber: parseInt(params.buildNumber, 10)});
+    const moduleBuildsPromise = new Resource({
+      url: `${config.apiRoot}/branches/builds/${repoBuild.id}/modules`,
+      type: 'GET'
+    }).send();
+    const moduleInfoPromise = _fetchModuleNames(params);
 
-    this.buildsPoller.disconnect();
-  }
-  
-  cancelBuild(cb) {
+    Q.spread([moduleInfoPromise, moduleBuildsPromise],
+      (moduleInfos, moduleBuilds) => {
+        const moduleBuildsWithNames = map(moduleBuilds, (build) => {
+          const moduleInfo = findWhere(moduleInfos, {id: build.moduleId});
+          const moduleInfoExtended = {
+            name: moduleInfo.name,
+            blazarPath: `${config.appRoot}/builds/branch/${params.branchId}/build/${params.buildNumber}/module/${moduleInfo.name}`
+          };
+
+          return extend(build, moduleInfoExtended);
+        });
+
+        cb(moduleBuildsWithNames);
+      });
+  });
+}
+
+function cancelBuild(params) {
+  return _fetchBranchBuildHistory(params).then((resp) => {
+    const repoBuild = findWhere(resp, {buildNumber: parseInt(params.buildNumber, 10)});
     const cancelPromise = new Resource({
-      url: `${config.apiRoot}/branches/builds/${this.buildsPoller.repoBuildId}/cancel`,
+      url: `${config.apiRoot}/branches/builds/${repoBuild.id}/cancel`,
       type: 'POST'
     }).send();
 
     cancelPromise.error((error) => {
-      console.warn(error);
-      cb(`Error cancelling build. See your console for more detail.`);
+      console.warn(error); // TODO: be better
     });
-       
-  }
-
-  getMalformedFiles(cb) {
-    if (this.branchId === undefined) {
-      setTimeout(() => {
-        this.getMalformedFiles(cb);
-      }, 500);
-
-      return;
-    }
-
-    const malformedFiles = new Resource({
-      url: `${config.apiRoot}/branches/${this.branchId}/malformedFiles`,
-      type: 'GET'
-    }).send();
-
-    malformedFiles.then((resp) => {
-      cb(resp);
-    });
-  }
+  }); 
 }
 
-export default RepoBuildApi;
+export default {
+  fetchRepoBuild: fetchRepoBuild,
+  fetchModuleBuilds: fetchModuleBuilds,
+  cancelBuild: cancelBuild
+};
